Update client to not require a successful ping
Please refer to https://github.com/gdiazlo/statsd as a solution to this issue in the original repo https://github.com/alexcesaro/statsd/issues/6.

This will add the functionality needed while keeping the additional changes that are currently present in this fork.

It is useful to have this change as the metrics within a service *shouldn't* stop a service from operating. Currently, it works as is but as soon as `.Clone` is called on the client, it crashes due to a nil conn (caused by this check).